Make the ui a bit better
Add a hover effect with zoom on the account tab along with a red hover effect for the Un-authenticate Account button of the connected account.

Add a zoom/opacity hover effect to all classic buttons.

Add the element-hover and element-hover-2 classes for reusable hover effects. The 2nd is because I encountered an error on the launcher tab in settings; this fixes the padding in that situation.